sched/fair: add support to compute perf deltas
When SchedTune is enabled, a CPU candidate for a boosted task can be one
which increase the overall energy consumption in order to improve its
performance (e.g. to reduce the time-to-completion for its activations).
Thus, whether a CPU candidate is suitable or not cannot be evaluated by
considering just the energy difference with respect to its previous CPU.
Instead, we need to estimate and evaluate a trade-off between energy
consumptions and performance.

This patch introduces the required support to compute a performance
index for a boosted task. This performance index has two main
contribution:

.:: SpeedUp_Index
=================
Estimated speed-up/slow-down in execution time due to the task being
executed at a different OPPs. When running on an higher OPP a task
activation should complete in less time. To the contrary, by running on
a lower OPP a task activation should require more time.

	SpeedUp_idx = SUI = cpu_boosted_capacity − task_util

Thus, this metric is mainly a function of the OPP of the CPU in which
the task will run.

NOTEs:
- We do not want to include any latency metric in this index, it
  should only try to estimate “how fast” a task will be executed
  in a given CPU.
- All the slow-down effects due to the placement of a task on a
  CPU will be considered by the delay index introduced in the
  following section

.:: Delay_Index
===============
Estimated delay for the completion of the current task’s activation due
to factors like the task being co-scheduled with other tasks. When the
task is scheduled alone on a CPU, then it will not experience delays. To
the contrary, if a task is co-scheduled with others, we can expect a
worst case delay which is when all other tasks runs before us.

       Delay_idx = DLI = 1024 * cpu_util / (task_util + cpu_util)
         (cpu_util does NOT include blocked load of the task)

Which is essentially (currently only) a measure of the CPU bandwidth
consumed by other tasks assuming they all run before us.

NOTEs:
- Smaller tasks gets more penalized than bigger tasks
  cpu_util includes the blocked load, which is ok since it represents a
  risk for potential delays due to tasks waking up on the same CPU in the
  near future
- cpu_util always discounts the current task (decayed) utilization, that
  way the cpu_util for both the src and dst CPUs always measure the
  utilization of that CPU without considering the current task
- cpu_util is the non boosted utilization of the CPU, otherwise we
  consider to be delayed also a boosted task which is running alone on a
  CPU
- cpu_util does not account (when PELT is in use) the bandwidth consumed
  by higher priority classes
- currently we do not account the task's FAIR PRIO. A task with higher
  FAIR PRIO usually gets longer slices at each scheduler delay period,
  thus this should mitigate the delay index for this task

.:: Performance_Index
=====================
Considering the previous two new metrics, the performance index is:

                   Perf_idx = SpeedUp_idx − Delay_idx

Which means that the new performance index:
- Estimate of "how fast" the task will run
- Discount all the latency treats

Change-Id: Ibac71dfff7eac689d3ca372a1b2470ce1619a8aa
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
(cherry picked from commit f4a9782b725f8210241b2f4b9fa9796d1c091825)