sched/{core,fair}: add support for estimated CPU utilization
Currently we have a cpu_util_freq function which has been added mainly
to report CPU utilization which are suitable to drive an OPP selection.

Let's extend that function to integrate UTIL_EST support, thus renaming
it to cpu_util_est which align it with the dual task_util_est.
When UTIL_EST support is enabled, and PELT is in use, the estimated
utilization for a CPU is returned.

The estimated utilization of a CPU is defined to be the maximum between the
PELT utilization and the sum(task_util_est) for all the currently RUNNABLE
tasks on that CPU.

This new definition of cpu_util_est biases OPP selections since schedutil
already uses the boosted_cpu_util, which is now driven by cpu_util_est.

Change-Id: Ia042a49c682621f7b2c2219870e5f1d5577ea40e
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
(cherry picked from commit d1746709cac2466ae23708054717e9ce8fef3039)